requirements.txt
selenium==3.141.0
PyGithub==1.54.1